chore(api): move chat validation tests out of generated code
Castiron-Internal-PR: https://github.com/openai/openai-python-internal/pull/52
Castiron-Source-SHA: c2505fd03faf7e816984054f585f34015c43df82
Castiron-Public-Base-SHA: 351ef84dc25211ea871f60eb16a469a89d20e133
